requirements.txt
discord.py[voice]==2.7.1
discord-ext-voice-recv @ git+https://github.com/vocolboy/discord-ext-voice-recv.git@main
python-dotenv==1.0.1
google-genai>=1.69.0
SpeechRecognition==3.14.1
